Remove more duplicated interface/abstract class
